Fixed tests

+++ src/test/java/ru/practicum/shareit/item/controller/ItemControllerTest.java
@@ -39,8 +39,8 @@ public class ItemControllerTest {
     @Autowired
     private ObjectMapper objectMapper;
 
-    private static final long USER_ID = 1;
-    private static final long ITEM_ID = 2;
+    private static final long USER_ID = 1L;
+    private static final long ITEM_ID = 2L;
 
     @SneakyThrows
     @Test
@@ -53,7 +53,7 @@ public class ItemControllerTest {
 
         mockMvc.perform(get("/items/" + ITEM_ID).header(HEADER_USER_ID, USER_ID))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id", is((int) ITEM_ID)))
+                .andExpect(jsonPath("$.id", is(ITEM_ID)))
                 .andExpect(jsonPath("$.name", is("Test item")));
     }
 
@@ -65,7 +65,7 @@ public class ItemControllerTest {
         item1.setName("Test item 1");
 
         ItemDto item2 = new ItemDto();
-        item2.setId(ITEM_ID + 1);
+        item2.setId(ITEM_ID + 1L);
         item2.setName("Test item 2");
 
         List<ItemDto> items = Arrays.asList(item1, item2);
@@ -74,9 +74,9 @@ public class ItemControllerTest {
 
         mockMvc.perform(get("/items").header(HEADER_USER_ID, USER_ID))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].id", is((int) ITEM_ID)))
+                .andExpect(jsonPath("$[0].id", is(ITEM_ID)))
                 .andExpect(jsonPath("$[0].name", is("Test item 1")))
-                .andExpect(jsonPath("$[1].id", is((int) ITEM_ID + 1)))
+                .andExpect(jsonPath("$[1].id", is( ITEM_ID + 1L)))
                 .andExpect(jsonPath("$[1].name", is("Test item 2")));
     }
 
@@ -88,7 +88,7 @@ public class ItemControllerTest {
         item1.setName("Test item 1");
 
         ItemDto item2 = new ItemDto();
-        item2.setId(ITEM_ID + 1);
+        item2.setId(ITEM_ID + 1L);
         item2.setName("Test item 2");
 
         List<ItemDto> items = Arrays.asList(item1, item2);
@@ -97,9 +97,9 @@ public class ItemControllerTest {
 
         mockMvc.perform(get("/items/search").param("text", "test"))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].id", is((int) ITEM_ID)))
+                .andExpect(jsonPath("$[0].id", is(ITEM_ID)))
                 .andExpect(jsonPath("$[0].name", is("Test item 1")))
-                .andExpect(jsonPath("$[1].id", is((int) ITEM_ID + 1)))
+                .andExpect(jsonPath("$[1].id", is( ITEM_ID + 1L)))
                 .andExpect(jsonPath("$[1].name", is("Test item 2")));
     }
